+++ vhost-device-console/src/main.rs
@@ -54,11 +54,16 @@ struct ConsoleArgs {
     #[clap(short = 's', long, value_name = "SOCKET")]
     socket_path: PathBuf,
 
+    /// Virtual machine communication endpoint.
+    /// Unix domain socket path (e.g., "/tmp/vm.sock").
+    #[clap(long, required(false), value_name = "VM_SOCKET")]
+    uds_path: Option<PathBuf>,
+
     /// Number of guests (sockets) to connect to.
     #[clap(short = 'c', long, default_value_t = 1)]
     socket_count: u32,
 
-    /// Console backend (Network, Nested) to be used.
+    /// Console backend (Network, Nested, Uds) to be used.
     #[clap(short = 'b', long, value_enum, default_value = "nested")]
     backend: BackendType,
 
@@ -87,14 +92,30 @@ impl TryFrom<ConsoleArgs> for VuConsoleConfig {
 
         let ConsoleArgs {
             socket_path,
+            uds_path,
             backend,
             tcp_port,
             socket_count,
             max_queue_size,
         } = args;
 
+        // check validation of uds_path under Uds mode.
+        if backend == BackendType::Uds {
+            let path = uds_path
+                .as_ref()
+                .filter(|p| !p.as_os_str().is_empty())
+                .ok_or(Error::InvalidUdsFile)?;
+
+            if let Some(parent_dir) = path.parent() {
+                if !parent_dir.exists() {
+                    return Err(Error::InvalidUdsFile);
+                }
+            }
+        }
+
         Ok(Self {
             socket_path,
+            uds_path: uds_path.unwrap_or_default(),
             backend,
             tcp_port,
             socket_count,