+++ vhost-device-console/CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 
 - [#811](https://github.com/rust-vmm/vhost-device/pull/811) Being able to specify max queue size
+- [#821](https://github.com/rust-vmm/vhost-device/pull/821) Add Unix domain socket backend support
 
 ### Changed